Dev: added database options to the export and import settings process
The export file now includes the database options set by WP Super Cache. This ensures that all settings are migrated over. The options are listed as a class variable and can be altered if necessary. The import process now takes these options into account and updates each individual option in the database with the imported value.