Bug: variables starting with the same string could replace multiple lines
The  is a good example. When this variable is passed through the wp_cache_replace_line function it will match and replace all variables starting with wp_cache_mobile leading to multiple declarations of the the  variable and removing the matched variables. Since  is not used in the wp_cache_replace_function in the plugin this bug is not usually seen.