Divide cache stats by 2 when regenerating them, for API.
If compression is enabled the plugin creates files called index.html.gz
for serving compressed data quickly. We took the decision to count
index.html and index.html.gz as one file on the cache contents page
but the code to divide the stats in half was in the display function,
not regenerate stats function. This moves that code so the REST API gets
the "correct" figure.
ref:
https://horizonfeedback.wordpress.com/2017/08/17/call-for-testing-calypso-wp-super-cache-extension/comment-page-1/#comment-399