Update feeds when site updates
Store a record of archive feed pages that can be deleted when the cache
is updated. This will help keep feeds fresh. Do not record feeds for
single pages. A maximum of 50 feeds is recorded to avoid bloating the
options table.
Will avoid writing to the config file and causing race issues reading
and writing.