Post Update now clears category cache
If a post was updated, it would clear the homepage and post cache, but the category pages, i.e. category/stories/... wouldn't clear. The function was set to only update categories when a post was changed from un-published to published or vice versa.
It's been changed so it clears the category cache when a post goes from un-published to published or a change is made when it's published.